compile

+++ src/qt/editaliasdialog.cpp
@@ -42,7 +42,7 @@ EditAliasDialog::EditAliasDialog(Mode mode, QWidget *parent) :
 	ui->expiryEdit->addItem(tr("5 Years"),QVariant(dateTime.toTime_t()));
 	// TODO for testnet
 	ui->expireTimeEdit->setText(QString::number(QDateTime::currentDateTimeUtc().addSecs(3600).toTime_t()));
-	//ui->expireTimeEdit->setText(ui->expiryEdit->itemData(0).toString());
+	//ui->expireTimeEdit->setText(QString::number(ui->expiryEdit->itemData(0).toInt()));
 	ui->expireTimeEdit->setEnabled(false);
 
     ui->privateDisclaimer->setText(QString("<font color='blue'>") + tr("This is to private profile information which is encrypted and only available to you. This is useful for when sending notes to a merchant through the payment screen so you don't have to type it out everytime.") + QString("</font>"));
@@ -108,7 +108,7 @@ void EditAliasDialog::onCustomExpireCheckBoxChanged(bool toggled)
 }
 void EditAliasDialog::expiryChanged(const QString& alias)
 {
-	ui->expireTimeEdit->setText(QString::number(ui->expiryEdit->itemData(ui->expiryEdit->currentIndex())));
+	ui->expireTimeEdit->setText(QString::number(ui->expiryEdit->itemData(ui->expiryEdit->currentIndex()).toInt()));
 }
 void EditAliasDialog::reqSigsChanged()
 {
